+++ scripts/download_and_examine.py
@@ -8,7 +8,7 @@ import hashlib
 import os
 from pathlib import Path
 
-from microscopy_portfolio import PortfolioManager
+from careamics_portfolio import PortfolioManager
 
 if __name__ == "__main__":
     # Create a portfolio object
