+++ src/careamics_portfolio/datasets/registry.txt
