+++ src/careamics_portfolio/portfolio_entry.py
