+++ src/careamics_portfolio/utils/pale_blue_dot.py
