+++ tests/utils.py
@@ -4,8 +4,8 @@ from typing import Dict, List
 
 import pytest
 
-from microscopy_portfolio.portfolio import IterablePortfolio
-from microscopy_portfolio.portfolio_entry import PortfolioEntry
+from careamics_portfolio.portfolio import IterablePortfolio
+from careamics_portfolio.portfolio_entry import PortfolioEntry
 
 
 def file_checker(path: Path, root_name: str, files: Dict[str, List[str]]) -> None:
